refactor(criteria): split up consolidated criteria files into single files
#32
