+++ src/criteria/criteria.ts
@@ -1,0 +1,3 @@
+import { Criterion } from "./criterion";
+
+export type Criteria<T = any> = Criterion<T>[];
