+++ src/criteria/index.ts
@@ -1,4 +1,10 @@
-export * from "./object-criteria";
-export * from "./object-criterion";
 export * from "./value-criterion";
 export * from "./values-criterion";
+
+export * from "./filter-by-criteria";
+export * from "./filter-by-criterion";
+export * from "./criteria";
+export * from "./criterion";
+export * from "./property-criteria";
+export * from "./reduce-criteria";
+export * from "./reduce-criterion";
