--- src/criteria/object-criteria.ts
@@ -1,57 +1,0 @@
-import { StringIndexable } from "../utils";
-import { ObjectCriterion } from "./object-criterion";
-
-export type ObjectCriteria<T = any> = ObjectCriterion<T>[];
-
-export module ObjectCriteria {
-    export function filter<T extends StringIndexable>(instances: T[], criteria: ObjectCriteria): T[] {
-        if (criteria.length === 0) {
-            return instances;
-        }
-
-        const allFiltered = new Set<T>();
-
-        for (const criterion of criteria) {
-            const filtered = ObjectCriterion.filter(instances, criterion);
-
-            for (const instance of filtered) {
-                allFiltered.add(instance);
-            }
-        }
-
-        return Array.from(allFiltered.values());
-    }
-
-    export function reduce(a: ObjectCriteria, b: ObjectCriteria): ObjectCriteria | null {
-        if (a.length === 0 && b.length === 0) {
-            return null;
-        }
-
-        let reduced = a.slice();
-        let didReduce = false;
-
-        for (let criterionB of b) {
-            let nextReduced: ObjectCriteria = [];
-
-            for (let criterionA of reduced) {
-                let reducedCriterion = ObjectCriterion.reduce(criterionB, criterionA);
-
-                if (reducedCriterion !== null) {
-                    nextReduced.push(reducedCriterion);
-                }
-
-                if (reducedCriterion !== criterionA && !didReduce) {
-                    didReduce = true;
-                }
-            }
-
-            reduced = nextReduced;
-        }
-
-        if (didReduce) {
-            return reduced.length > 0 ? reduced : null;
-        } else {
-            return a;
-        }
-    }
-}
