+++ src/query/query.ts
@@ -1,10 +1,10 @@
 import { Class } from "../utils";
 import { Selection } from "../selection";
-import { ObjectCriteria } from "../criteria";
+import { Criteria } from "../criteria";
 import { Reducible } from "./reducible";
 
 export interface Query {
-    criteria: ObjectCriteria;
+    criteria: Criteria;
     model: Class[];
     options: Reducible;
     selection: Selection;
